+++ apps/backend/package.json
@@ -3,7 +3,6 @@
   "module": "index.ts",
   "type": "module",
   "version": "0.0.1",
-
   "scripts": {
     "lint": "biome check .",
     "format": "biome format --write .",
@@ -20,7 +19,6 @@
     "check-types": "tsc --noEmit"
   },
   "dependencies": {
-    "@director.run/store": "workspace:*",
     "@modelcontextprotocol/sdk": "^1.8.0",
     "@trpc/server": "^11.0.2",
     "cli-table3": "^0.6.5",
@@ -30,6 +28,7 @@
     "lodash": "^4.17.21",
     "pino": "^9.6.0",
     "pino-pretty": "^13.0.0",
+    "slugify": "^1.6.6",
     "superjson": "^2.2.1",
     "zod": "^3.24.2"
   },
@@ -41,6 +40,7 @@
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
     "@types/lodash": "^4.17.16",
-    "@types/semver": "^7.7.0"
+    "@types/semver": "^7.7.0",
+    "type-fest": "^4.39.1"
   }
 }
