+++ apps/backend/src/commands/seed.ts
@@ -1,14 +1,12 @@
-import { writeStore } from "@director.run/store";
-import type { Config } from "@director.run/store/schema";
-import { PROXY_DB_FILE_PATH } from "../config";
+import { writeStore } from "../config";
+import type { Config } from "../config/schema";
+import { PROXY_DB_FILE_PATH } from "../constants";
 import { getLogger } from "../helpers/logger";
 
 export async function seed() {
   const logger = getLogger("seed");
   logger.info(`Seeding database at path: ${PROXY_DB_FILE_PATH}`);
   const seedProxyDB: Config = {
-    version: "beta",
-    port: 3000,
     proxies: [
       {
         id: "my-first-proxy",
