+++ apps/backend/src/helpers/read-json.ts
