--- apps/backend/turbo.json
@@ -1,14 +1,0 @@
-{
-  "extends": ["//"],
-  "tasks": {
-    "build": {
-      "dependsOn": ["^@director.run/store#build"]
-    },
-    "dev": {
-      "dependsOn": ["^@director.run/store#build"]
-    },
-    "test": {
-      "dependsOn": ["^@director.run/store#build"]
-    }
-  }
-}
