+++ apps/desktop/src/components/connection/connection-provider.tsx
@@ -7,7 +7,7 @@ import { useTimeout } from "usehooks-ts";
 import { assertUnreachable } from "@/lib/assert-unreachable";
 import { createCtx } from "@/lib/create-ctx";
 import { trpc } from "@/lib/trpc/trpc";
-import { Proxy } from "@director.run/store/schema";
+import { Proxy } from "@director/backend/src/config/schema";
 import { Loader2Icon } from "lucide-react";
 import { useLocation, useNavigate } from "react-router";
 import { GlobalLayout } from "../global-layout";
