--- packages/store/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@director.run/store",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "main": "./dist/index.js",
-  "exports": {
-    "./package.json": "./package.json",
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    },
-    "./schema": {
-      "types": "./dist/schema.d.ts",
-      "import": "./dist/schema.js",
-      "require": "./dist/schema.cjs"
-    },
-    "./constants": {
-      "types": "./dist/constants.d.ts",
-      "import": "./dist/constants.js",
-      "require": "./dist/constants.cjs"
-    }
-  },
-  "scripts": {
-    "build": "tsup",
-    "check-types": "tsc --noEmit",
-    "clean": "rm -rf dist .turbo node_modules",
-    "dev": "bun run --watch ./src/index.ts",
-    "format": "biome format --write .",
-    "lint": "biome check .",
-    "test": "vitest"
-  },
-  "dependencies": {
-    "slugify": "^1.6.6",
-    "zod": "^3.24.2"
-  },
-  "devDependencies": {
-    "@director.run/tsconfig": "workspace:*",
-    "tsup": "^8.4.0",
-    "type-fest": "^4.39.1",
-    "vitest": "^3.1.1"
-  }
-}
